refactor: split project in multiple files
