+++ src/constants.h
@@ -29,7 +29,6 @@
 
 #define GAME_SPEED 200.0
 #define ANIMATION_SPEED 14 // BIGGER IS SLOWER
-#define SPRITE_SPEED 1     // Don't change this for the moment
 
 #define GHOST_INITIAL_POS_X 4
 #define GHOST_INITIAL_POS_Y 124
@@ -42,8 +41,6 @@
 #define PACMAN_SPRITE_MOUTHS_DIRECTION 3
 #define PACMAN_SPACING_X 3
 #define PACMAN_SPACING_Y 2
-#define PACMAN_ARROW_SIZE 7
-#define PACMAN_ARROW_SPACING 20
 
 #define PACMAN_DEATH_ANIMATION_FRAMES 10
 #define PACMAN_DEATH_ANIMATION_DURATION 200
