+++ src/enums/direction.h
@@ -3,6 +3,8 @@
 
 #define DIRECTION_COUNT 4
 
+#include <stdlib.h>
+
 typedef enum
 {
     DIRECTION_RIGHT = 0,
@@ -10,5 +12,11 @@ typedef enum
     DIRECTION_UP = 2,
     DIRECTION_DOWN = 3,
 } Direction;
+
+Direction getNextDirection(Direction direction);
+
+Direction getRandomDirection();
+
+Direction getOppositeDirection(Direction direction);
 
 #endif
