--- src/enums/ghost-type.h
@@ -1,16 +1,0 @@
-//
-// Created by Franck GUTMANN on 23/02/2023.
-//
-
-#ifndef PACMAN_GHOST_TYPE_ENUM_H
-#define PACMAN_GHOST_TYPE_ENUM_H
-
-typedef enum
-{
-    RED = 0,
-    PINK = 1,
-    BLUE = 2,
-    ORANGE = 3,
-} GhostType;
-
-#endif //PACMAN_GHOST_TYPE_ENUM_H
