+++ src/ui/game_window.h
@@ -13,13 +13,15 @@
 #include "variables.h"
 #include "models/pacman.h"
 #include "models/position.h"
-#include "models/ghost.h"
+#include "models/ghosts/ghost.h"
 #include "models/game.h"
 #include "ui/game_info_panel.h"
 #include "ui/menu_window.h"
 #include "ui/pause_menu.h"
 #include "models/fruit.h"
 #include "models/audio.h"
+#include "models/coins.h"
+#include "models/arrow.h"
 
 #define READY_W 46
 #define READY_H 7
