+++ src/utilities/array_utils.h
@@ -59,4 +59,8 @@ char get2DArrayElement(char **array, int row, int column);
  */
 struct Position find2DArrayElement(char **array, int rows, int columns, char element);
 
+int countArrayElementEqualsTo(int *array, int size, int value);
+
+int findIndexOf(int *array, int size, int value);
+
 #endif //PACMAN_ARRAY_UTILS_H
